requirements/dev.txt
-r prod.txt
black==20.8b1
django-factory-boy==1.0.0
flake8==3.8.4
mypy==0.790
pytest==6.2.1
pytest-django==4.1.0
Sphinx==3.4.3
pytest-cov==2.11.1
pytest-mock==3.5.1
pytest-env==0.6.2
